api/requirements.txt
# ZUGFeRD generation
drafthorse>=2.3.0
factur-x>=3.15
fpdf2>=2.8.0
lxml>=4.9.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# File parsing
pdfplumber>=0.11.0
pandas>=2.2.0
openpyxl>=3.1.0
Pillow>=10.0.0
pillow-heif>=0.18.0

# AI fallback
openai>=1.0.0
tenacity>=8.0.0

# Utilities
aiofiles>=23.0.0

# Dev
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
